backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
pandas==2.0.3
numpy==1.26.4
pydantic==2.9.0
beautifulsoup4==4.12.3
feedparser==6.0.11
textblob==0.18.0

